Desktop/Ai_assistant/requirements.txt
python-dotenv==1.0.1
elevenlabs==0.4.3